@@ -2,14 +2,13 @@ package ar.com.plug.examen.domain.service;
 
 import ar.com.plug.examen.domain.model.Message;
 import ar.com.plug.examen.domain.service.impl.ProcessMessageServiceImpl;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
-import org.mockito.junit.MockitoJUnitRunner;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@RunWith(MockitoJUnitRunner.class)
 public class ProcessMessageServiceTest {
 
     @InjectMocks
@@ -21,7 +20,7 @@ public class ProcessMessageServiceTest {
         String messageTest = "TEST";
         Message message = messageService.processMessage(messageTest);
 
-        assertNotNull(message);
+        Assertions.assertNotNull(message);
         assertEquals(message.getMessage(), messageTest);
     }
 }
